Move pkg/radrp/handlers to pkg/handlers
