+++ pkg/handlers/azure_cosmosdb_base.go
