+++ pkg/workloads/daprpubsubv1alpha1/render.go
@@ -11,8 +11,8 @@ import (
 	"fmt"
 
 	"github.com/Azure/radius/pkg/azresources"
+	"github.com/Azure/radius/pkg/handlers"
 	"github.com/Azure/radius/pkg/model/components"
-	"github.com/Azure/radius/pkg/radrp/handlers"
 	"github.com/Azure/radius/pkg/radrp/outputresource"
 	"github.com/Azure/radius/pkg/workloads"
 )
