+++ pkg/workloads/daprpubsubv1alpha1/render_test.go
@@ -9,8 +9,8 @@ import (
 	"context"
 	"testing"
 
+	"github.com/Azure/radius/pkg/handlers"
 	"github.com/Azure/radius/pkg/model/components"
-	"github.com/Azure/radius/pkg/radrp/handlers"
 	"github.com/Azure/radius/pkg/radrp/outputresource"
 	"github.com/Azure/radius/pkg/workloads"
 	"github.com/stretchr/testify/require"
