+++ pkg/workloads/daprstatestorev1alpha1/sql.go
@@ -8,7 +8,7 @@ package daprstatestorev1alpha1
 import (
 	"errors"
 
-	"github.com/Azure/radius/pkg/radrp/handlers"
+	"github.com/Azure/radius/pkg/handlers"
 	"github.com/Azure/radius/pkg/radrp/outputresource"
 	"github.com/Azure/radius/pkg/workloads"
 )
